requirements.txt
alembic==1.18.4
aniso8601==10.0.1
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
click==8.3.1
cloudinary==1.44.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
eventlet==0.40.4
Faker==18.9.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.61.1
greenlet==3.3.2
gunicorn==25.1.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2025.2
setuptools==82.0.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.47
sqlalchemy-serializer==1.6.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wsproto==1.3.2
WTForms==3.2.1
